third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-anchor-cross-origin.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<a id="a" href="https://does-not-exist/foo.html"></a>
<script>
async_test(t => {
  appHistory.onnavigate = t.step_func_done(e => {
    assert_true(e.cancelable);
    assert_false(e.canRespond);
    assert_false(e.userInitiated);
    assert_false(e.hashChange);
    assert_equals(e.formData, null);
    assert_equals(e.destination.url, "https://does-not-exist/foo.html");
    assert_false(e.destination.sameDocument);
    e.preventDefault();
  });
  a.click();
}, "<a> cross-origin navigate event");
</script>
